refactor: move AVL struct in general render into its renderer
